# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Drachenboot Manager API",
        "version": "2.5.0",
        "description": "API documentation for the Drachenboot Manager application with multi-team support."
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/events/{id}/assignments": {
            "get": {
                "summary": "Get Assignments",
                "tags": [
                    "Assignments"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of assignments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Assignment"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "summary": "Update Assignment",
                "tags": [
                    "Assignments"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "seatId": {
                                        "type": "string"
                                    },
                                    "paddlerId": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "isCanister": {
                                        "type": "boolean"
                                    }
                                },
                                "required": [
                                    "seatId"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Assignment updated"
                    }
                }
            }
        },
        "/api/events/{id}/attendance": {
            "post": {
                "summary": "Set Attendance",
                "tags": [
                    "Attendance"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "paddlerId": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "yes",
                                            "no",
                                            "maybe"
                                        ]
                                    }
                                },
                                "required": [
                                    "paddlerId",
                                    "status"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Attendance updated"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Assignment": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "eventId": {
                        "type": "string"
                    },
                    "seatId": {
                        "type": "string",
                        "description": "Seat identifier (e.g., 'row-1-left', 'drummer', 'steer')"
                    },
                    "paddlerId": {
                        "type": "string",
                        "nullable": true
                    },
                    "isCanister": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "id",
                    "eventId",
                    "seatId",
                    "isCanister"
                ]
            },
...
        },
        "securitySchemes": {
            "sessionAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "authjs.session-token",
                "description": "Session-based authentication using NextAuth.js. Requires active login session."
            }
        }
    },
    "security": [
        {
            "sessionAuth": []
        }
    ],
...
}